requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.9
psycopg2>=2.9.6
alembic>=1.10.3

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=1.10.7
pydantic-settings>=2.0.0
python-dateutil>=2.8.2

# AWS S3 for optional file storage
boto3>=1.26.115

# Email
jinja2>=3.1.2
aiosmtplib>=2.0.1

# XML handling for tax returns
defusedxml>=0.7.1

# Utilities
python-dotenv>=1.0.0
email-validator>=2.0.0
httpx>=0.24.0

boto3
PyJWT>=2.6.0